core code updates - enhancements
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@6159 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/dbinfo/DatabaseInformationFull.java
@@ -1597,7 +1597,7 @@ extends org.hsqldb.dbinfo.DatabaseInformationMain {
                             if (tableIdList[j] == spaceId) {
                                 allocated += fileBlockSize;
                                 used += fileBlockSize
-                                        - freeSpaceList[j] * cacheScale;
+                                        - (long) freeSpaceList[j] * cacheScale;
                             }
                         }
                     }
@@ -1631,7 +1631,7 @@ extends org.hsqldb.dbinfo.DatabaseInformationMain {
             for (int j = 0; j < tableIdList.length; j++) {
                 if (tableIdList[j] == DataSpaceManager.tableIdDefault) {
                     allocated += fileBlockSize;
-                    used      += fileBlockSize - freeSpaceList[j] * cacheScale;
+                    used      += fileBlockSize - (long) freeSpaceList[j] * cacheScale;
                 } else if (tableIdList[j] == DataSpaceManager.tableIdEmpty
                            && bitMapList[j] != 0) {
                     empty += fileBlockSize;
