+++ trunk/src/org/hsqldb/server/OdbcPacketInputStream.java
@@ -205,7 +205,7 @@ class OdbcPacketInputStream extends DataInputStream {
 
         try {
             return (len < 0) ? null
-                             : new BinaryData((long) len, this);
+                             : new BinaryData(len, this);
         } catch (HsqlException he) {
             throw new IOException(he.getMessage());
         }
